+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add subcommand 'check', which is the original command from 0.0.1
 - Add help messages for main command and `check`
 - Support option `--ignore-throws`, which makes DrString ignore throws in docstring validation.
+- Colored output for TTY
 
 ## 0.0.1